udp cookie auth

+++ src/main/java/com/example/autoluxe/domain/User.java
@@ -10,6 +10,7 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.Collections;
@@ -53,7 +54,7 @@ public class User implements UserDetails, Serializable {
     private String epic_token;
 
     @Column(name = "balance")
-    private Double balance;
+    private BigDecimal balance;
 
     @ToString.Exclude
     @OneToMany(mappedBy = "user",orphanRemoval = true)
